Split verification methods from VectorStoreRecordPropertyReader
